requirements.txt
flask
tqdm
coloredlogs
colorama
markovify
Mastodon.py
tweepy
pony
todoist-python